Set up Phing build tool to coordinate jobs in commit stage
This reduces the coupling to CircleCi and makes it easy to run the same
set of jobs in local.

I haven't used Phing for years and felt like trying it out again.

Not 100% convinced yet - I don't like having to set passthru="true" and checkreturn="true"
for each exec job - I feel like these should be defaults. And it seems
that passing arguments to commands will be more verbose than necassary.
May switch to a tool that isn't based on XML later.

Also of course since phing is installed with composer you can't use
phing to run the composer build.